챌린지 식별값 타입 변경 (#37)
* refactor: id int -> long 수정

* refactor: id int --> long 수정

* refactor: id int ---> long 수정

+++ src/main/java/sarangbang/site/challengemember/repository/ChallengeMemberRepository.java
@@ -5,5 +5,5 @@ import sarangbang.site.challengemember.entity.ChallengeMember;
 
 public interface ChallengeMemberRepository extends JpaRepository<ChallengeMember, Long> {
 
-    int countByChallengeId(int challengeId);
+    Long countByChallengeId(Long challengeId);
 }
